+++ Pandora/Paradigm.hs
@@ -1,7 +1,7 @@
 module Pandora.Paradigm (module Exports) where
 
 import Pandora.Paradigm.Structure as Exports
-import Pandora.Paradigm.Inventory as Exports
+import Pandora.Paradigm.Inventory.Some as Exports
 import Pandora.Paradigm.Controlflow as Exports
 import Pandora.Paradigm.Primary as Exports
 import Pandora.Paradigm.Schemes as Exports