+++ Pandora/Paradigm/Inventory/Ability/Modifiable.hs
@@ -1,8 +1,8 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
-module Pandora.Paradigm.Inventory.Ability.Modifiable where
+module Pandora.Paradigm.Inventory.Some.Ability.Modifiable where
 
-import Pandora.Paradigm.Inventory.Ability.Viewable (Viewable)
-import Pandora.Paradigm.Inventory.Ability.Replaceable (Replaceable)
+import Pandora.Paradigm.Inventory.Some.Ability.Viewable (Viewable)
+import Pandora.Paradigm.Inventory.Some.Ability.Replaceable (Replaceable)
 
 class (Viewable i, Replaceable i) => Modifiable i where
 	type Modification i l r :: *