+++ Pandora/Paradigm/Inventory/Some.hs
@@ -1,0 +1,9 @@
+module Pandora.Paradigm.Inventory.Some (module Exports) where
+
+import Pandora.Paradigm.Inventory.Some.Optics as Exports
+import Pandora.Paradigm.Inventory.Some.State as Exports
+import Pandora.Paradigm.Inventory.Some.Store as Exports
+import Pandora.Paradigm.Inventory.Some.Imprint as Exports
+import Pandora.Paradigm.Inventory.Some.Accumulator as Exports
+import Pandora.Paradigm.Inventory.Some.Provision as Exports
+import Pandora.Paradigm.Inventory.Some.Equipment as Exports