+++ Pandora/Paradigm/Structure/Ability/Accessible.hs
@@ -1,7 +1,7 @@
 module Pandora.Paradigm.Structure.Ability.Accessible where
 
 import Pandora.Paradigm.Primary.Functor.Identity (Identity)
-import Pandora.Paradigm.Inventory.Optics (Lens)
+import Pandora.Paradigm.Inventory.Some.Optics (Lens)
 
 class Accessible target source where
 	access :: Lens Identity source target