+++ Pandora/Paradigm/Structure/Ability/Possible.hs
@@ -1,7 +1,7 @@
 module Pandora.Paradigm.Structure.Ability.Possible where
 
 import Pandora.Paradigm.Primary.Functor.Maybe (Maybe)
-import Pandora.Paradigm.Inventory.Optics (Lens)
+import Pandora.Paradigm.Inventory.Some.Optics (Lens)
 
 class Possible target source where
 	perhaps :: Lens Maybe source target