+++ Pandora/Paradigm/Structure/Interface/Set.hs
@@ -19,7 +19,7 @@ import Pandora.Paradigm.Primary.Object.Boolean (Boolean)
 import Pandora.Paradigm.Primary.Object.Numerator (Numerator (Zero))
 import Pandora.Paradigm.Schemes.T_U (type (<:.:>))
 import Pandora.Paradigm.Structure.Ability.Morphable (Morphable (Morphing), Morph (Find), find)
-import Pandora.Paradigm.Inventory.State (State, modify)
+import Pandora.Paradigm.Inventory.Some.State (State, modify)
 import Pandora.Paradigm.Controlflow.Effect (run, (!))
 
 type Set t f a = (Traversable (->) (->) t, Setoid a, Setoid (t a), Morphable (Find f) t)